% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate}
\alias{generate}
\title{Generate bootstrap and permutation resamples}
\usage{
generate(x, n_samples, type = "bootstrap")
}
\arguments{
\item{x}{Tibble or data.frame, containing the response variable from which to sample}

\item{n_samples}{Integer. The number of samples to be generated. Cannot be negative or 0.}

\item{type}{String. Specifies the method to regenerate the samples. Default: "bootstrap" (the future implementations will allow for other types including "Permutation")}
}
\value{
Tibble containing all the resamples in a dataframe with an additional column mapping each observation to its sample id.
}
\description{
Takes a tibble and generates n samples using either the bootstrapping or permutation methods
}
\examples{
library(dplyr)
gap_df <- gapminder::gapminder \%>\%
    dplyr::filter(year == 2007 & continent \%in\% c("Africa"))
generate(gap_df, n_samples=5, type="bootstrap")
}
