% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ci.R
\name{get_ci}
\alias{get_ci}
\title{Get Confidence Intervals}
\usage{
get_ci(data, column = NULL, confidence_level = 0.95,
  point_estimate = NULL, type = "percentile")
}
\arguments{
\item{data}{A dataframe/tibble with the calculated statistics in a column named 'stat'}

\item{column}{A string that specifies the column name in the dataframe that will be used for calculations. Must be specified.}

\item{confidence_level}{The confidence level specified  as a float between 0 and 1 inclusive. Default is set at 0.95}

\item{point_estimate}{A value representing a calculated statistic of the original sample. Will be set to "NULL" if method is percentile.}

\item{type}{A string specifying the type of method to be used to calculate the confidence intervals. Options are "standarderror" or "percentile".
As of this version only percentile has been implemented. Default is "percentile".}
}
\value{
1 x 4 dataframe/tibble with numerical values representing the Confidence Level, Lower Bound, Upper Bound, and Point Estimate
}
\description{
Get Confidence Intervals
}
\examples{

df_t <- data.frame(runif(20))
names(df_t) <- c("randoms")
get_ci(data=df_t,column="randoms",confidence_level=0.9)

}
