% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.R
\name{specify}
\alias{specify}
\title{Specify the response from datr frame, and predictors (if any)}
\usage{
specify(x, response, explanatory = NULL)
}
\arguments{
\item{response}{One column which will be chosen as repsonse}

\item{explanatory}{The columns to be the explanatory variables}

\item{df}{A data frame}
}
\value{
Dataframe containing one column for response variable and zero or more columns for the explanatory variables. The first column is always the response.
}
\description{
Specify the response from datr frame, and predictors (if any)
}
\examples{
# select time data from data frame `lung`
specify(lung,"time")

}
